split program
