+++ src/qibosoq/programs/__init__.py
@@ -1,0 +1,1 @@
+"""Program modules."""
